Update Mega-Linter workflow for main, bandit, isort (#14)
Fix Mega-Linter workflow to refer to main. The Mega-Linter workflow has references to master, but the default branch was always main in this repo. This fixes conditions, updates the name in comments, and removes the branch matching rule.

Skip new bandit security warnings from Mega-Linter for subprocess and shell=True. Arbitrary code is what we intent to run. Ignore comment nosec needs to be on the line with shell=True to work.

Add isort config to align comments in imports with Black.

Add nosec bandit comment to known words.